requirements.txt
Flask==3.0.0
google-generativeai==0.8.6
protobuf==5.29.6
python-dotenv==1.0.0
requests==2.31.0
